Create FeedItemMenuBuilder to incapsulate logics with feed item popup menu creation, fix issue with wrong text in the feed menu items

+++ app/src/main/java/com/worldventures/dreamtrips/modules/feed/view/cell/base/FeedHeaderCell.java
@@ -3,8 +3,8 @@ package com.worldventures.dreamtrips.modules.feed.view.cell.base;
 import android.app.Dialog;
 import android.support.annotation.MenuRes;
 import android.support.annotation.StringRes;
-import android.support.v7.widget.PopupMenu;
 import android.view.View;
+import android.widget.ImageView;
 
 import com.innahema.collections.query.queriables.Queryable;
 import com.techery.spares.session.SessionHolder;
@@ -21,6 +21,7 @@ import com.worldventures.dreamtrips.modules.feed.event.ProfileClickedEvent;
 import com.worldventures.dreamtrips.modules.feed.model.FeedItem;
 import com.worldventures.dreamtrips.modules.feed.model.comment.Comment;
 import com.worldventures.dreamtrips.modules.feed.view.custom.FeedActionPanelView;
+import com.worldventures.dreamtrips.modules.feed.view.popup.FeedItemMenuBuilder;
 import com.worldventures.dreamtrips.modules.feed.view.util.CommentCellHelper;
 import com.worldventures.dreamtrips.modules.feed.view.util.FeedActionPanelViewActionHandler;
 import com.worldventures.dreamtrips.modules.feed.view.util.FeedItemHeaderHelper;
@@ -48,6 +49,8 @@ public abstract class FeedHeaderCell<T extends FeedItem> extends AbstractCell<T>
     View commentDivider;
     @InjectView(R.id.actionView)
     FeedActionPanelView actionView;
+    @InjectView(R.id.edit_feed_item)
+    ImageView editFeedItem;
 
     @Inject
     ActivityRouter activityRouter;
@@ -69,7 +72,7 @@ public abstract class FeedHeaderCell<T extends FeedItem> extends AbstractCell<T>
     @Override
     protected void syncUIStateWithModel() {
         feedItemHeaderHelper.set(getModelObject(), itemView.getContext(), sessionHolder.get().get().getUser().getId(), false);
-        feedItemHeaderHelper.setOnEditClickListener(this::onEditClicked);
+        feedItemHeaderHelper.setOnEditClickListener(view -> onMore());
         if (commentCellHelper != null) {
             Comment comment = getModelObject().getItem().getComments() == null ? null :
                     Queryable.from(getModelObject().getItem().getComments())
@@ -117,56 +120,30 @@ public abstract class FeedHeaderCell<T extends FeedItem> extends AbstractCell<T>
     }
 
     protected void showMoreDialog(@MenuRes int menuRes, @StringRes int headerDelete, @StringRes int textDelete) {
-        // temporally, while edit button doesn't exists
-        PopupMenu popup = new PopupMenu(itemView.getContext(), commentPreview);
-        popup.inflate(menuRes);
-        popup.setOnMenuItemClickListener(item -> {
-            switch (item.getItemId()) {
-                case R.id.action_delete:
-                    Dialog dialog = new SweetAlertDialog(itemView.getContext(), SweetAlertDialog.WARNING_TYPE)
-                            .setTitleText(itemView.getResources().getString(headerDelete))
-                            .setContentText(itemView.getResources().getString(textDelete))
-                            .setConfirmText(itemView.getResources().getString(R.string.post_delete_confirm))
-                            .setConfirmClickListener(sDialog -> {
-                                sDialog.dismissWithAnimation();
-                                onDelete();
-                            });
-                    dialog.setCanceledOnTouchOutside(true);
-                    dialog.show();
-                    break;
-                case R.id.action_edit:
-                    onEdit();
-                    break;
-            }
-
-            return true;
-        });
-        popup.show();
+        FeedItemMenuBuilder.create(itemView.getContext(), editFeedItem, menuRes)
+                .onDelete(() -> showDeleteDialog(headerDelete, textDelete))
+                .onEdit(this::onEdit)
+                .show();
     }
 
-    protected void onEditClicked(View v) {
-        PopupMenu popup = new PopupMenu(v.getContext(), v);
-        popup.inflate(R.menu.menu_feed_entity_edit);
-        popup.setOnMenuItemClickListener(item -> {
-            switch (item.getItemId()) {
-                case R.id.action_delete:
+    private void showDeleteDialog(@StringRes int headerDelete, @StringRes int textDelete) {
+        Dialog dialog = new SweetAlertDialog(itemView.getContext(), SweetAlertDialog.WARNING_TYPE)
+                .setTitleText(itemView.getResources().getString(headerDelete))
+                .setContentText(itemView.getResources().getString(textDelete))
+                .setConfirmText(itemView.getResources().getString(R.string.post_delete_confirm))
+                .setConfirmClickListener(sDialog -> {
+                    sDialog.dismissWithAnimation();
                     onDelete();
-                    break;
-                case R.id.action_edit:
-                    onEdit();
-                    break;
-            }
-
-            return true;
-        });
-        popup.show();
+                });
+        dialog.setCanceledOnTouchOutside(true);
+        dialog.show();
     }
 
     protected abstract void onDelete();
 
-    protected abstract void onMore();
-
     protected abstract void onEdit();
+
+    protected abstract void onMore();
 
     protected void openComments(FeedItem baseFeedModel) {
         openComments(baseFeedModel, false);
